+++ debug_blackbook.rb
@@ -4,7 +4,7 @@ aol = Blackbook.get :username => 'mondragonmichael@aol.com', :password => '123ao
 puts "===== aol #{aol.inspect}"
 
 unless aol.detect{|c| c[:name].downcase == 'Dave Myron'.downcase && c[:email] == 'dave.myron@contentfree.com'}
-    puts "AOL not working"
-    exit 1
-  end
+  puts "AOL not working"
+  exit 1
+end